+++ scripts/services/rules-assistance-service.mjs
@@ -74,6 +74,7 @@ export class RulesAssistanceService {
   }
 
   static async ready() {
+    await EffectLifecycleService.ready();
     await AgonizingBlastBindingService.ready();
     await MageArmorAssistanceService.ready();
     await BardicInspirationAssistanceService.ready();